project/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
psycopg2==2.9.9
PyJWT==2.9.0
pymongo==4.9.1
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.2
